[#273] Use AtomicUsize for maximum compatibility

+++ src/relay/flow.rs
@@ -3,7 +3,7 @@
 use std::{
     collections::BTreeMap,
     sync::{
-        atomic::{AtomicU64, Ordering},
+        atomic::{AtomicUsize, Ordering},
         Arc,
     },
 };
@@ -12,36 +12,36 @@ use crate::config::Config;
 
 /// Flow statistic for one server
 pub struct FlowStatistic {
-    tx: AtomicU64,
-    rx: AtomicU64,
+    tx: AtomicUsize,
+    rx: AtomicUsize,
 }
 
 impl FlowStatistic {
     /// Create an empty statistic
     pub fn new() -> FlowStatistic {
         FlowStatistic {
-            tx: AtomicU64::new(0),
-            rx: AtomicU64::new(0),
+            tx: AtomicUsize::new(0),
+            rx: AtomicUsize::new(0),
         }
     }
 
     /// Total bytes transferred
-    pub fn tx(&self) -> u64 {
+    pub fn tx(&self) -> usize {
         self.tx.load(Ordering::Acquire)
     }
 
     /// Add bytes transferred
-    pub fn incr_tx(&self, tx: u64) {
+    pub fn incr_tx(&self, tx: usize) {
         self.tx.fetch_add(tx, Ordering::AcqRel);
     }
 
     /// Total bytes received
-    pub fn rx(&self) -> u64 {
+    pub fn rx(&self) -> usize {
         self.rx.load(Ordering::Acquire)
     }
 
     /// Add bytes received
-    pub fn incr_rx(&self, rx: u64) {
+    pub fn incr_rx(&self, rx: usize) {
         self.rx.fetch_add(rx, Ordering::AcqRel);
     }
 }
@@ -86,7 +86,7 @@ impl ServerFlowStatistic {
     }
 
     /// Transmission statistic for manager
-    pub fn trans_stat(&self) -> u64 {
+    pub fn trans_stat(&self) -> usize {
         self.tcp().tx() + self.tcp().rx() + self.udp().tx() + self.udp.rx()
     }
 }
